zephyr: Set PIXIT parameters for the GATT test cases
This adds a function to be called to update appropriate PIXIT
parameters for the test cases.
It's helpful to have this, because PIXIT are set to default everytime
the TPG is updated. To avoid starting PTS in gui mode to update the
pixits, this can be done using this function. Then we're sure that
those are set to appropriate value.