btp: Change the data conversion gap_device_found_ev
This patch changes the conversion of the data in gap_device_found_ev.
Data passed to the gap_device_found_ev won't be converted to it's
binary representation. Instead of that, data received from BTP Device
Found event will be unpacked and address will be converted to
it's hexadecimal representation.