btp: Fix missing advertising reports
This fixes missing advertising reports from Zephyr.
The reason why those were missing was that Zephyr was sending Device Found
events right after the Discovery started. But on our side, we were waiting
for MMI from PTS to read form the BTP socket. In this case, when there were
many events sent, we failed to read every event.

With this patch advertisements are read immediately when discovery started
and put on a list for later verification.