testcase: Run post_wid TestFunc's right after sending MMI response
The thread is started to run post_wid commands right after responding
to the MMI. Thanks to that, those commands are not waiting for another
MMI to execute.

This is crucial in situation where PTS wants us to Start discovery and
turns off advertising when sends another MMI.
The discovery would fail in such situation.

 1/12   GAP   TC_BROB_OBSV_BV_01_C   PASS
 2/12   GAP   TC_BROB_OBSV_BV_02_C   PASS
 3/12   GAP   TC_DISC_LIMP_BV_01_C   PASS
 4/12   GAP   TC_DISC_LIMP_BV_02_C   PASS
 5/12   GAP   TC_DISC_LIMP_BV_03_C   PASS
 6/12   GAP   TC_DISC_LIMP_BV_04_C   PASS
 7/12   GAP   TC_DISC_LIMP_BV_05_C   PASS
 8/12   GAP   TC_DISC_GENP_BV_01_C   PASS
 9/12   GAP   TC_DISC_GENP_BV_02_C   PASS
10/12   GAP   TC_DISC_GENP_BV_03_C   PASS
11/12   GAP   TC_DISC_GENP_BV_04_C   PASS
12/12   GAP   TC_DISC_GENP_BV_05_C   PASS