zephyr: mesh_wid: Workaround failing MESH/NODE/IVU/BV-02-C
In this test case, PTS is waiting IUT to back to the Normal Mode,
while in IV Update Test mode. Disabling IVU test mode will cause
IUT to wait 96 hours to back to the normal state.
So do not leave test mode, and just give IUT time to go back
to the Normal operation mode.