mynewt: fix for L2CAP/LE/CFC/BV-06-C
Test L2CAP/LE/CFC/BV-06-C requires from us to send 4 frames of any
size. The credits allowing IUT to send them are coming in doses of 1, 1
and 2. We need to send data which will consume a credit, wait for PTS
to register it and send another credit, and repeat it 4 times total.