zephyr: Workaround for issue with custom config file
Use '-DOVERLAY_CONFIG' option to specify custom settings that will be
merged with the default config. No need for copying and editing the
default configuration anymore.

'-DCONF_FILE' and '-DOVERLAY_CONFIG' seem to do the same thing now and
using a copy of the default config with those options is broken so this
is a workaround for that issue.