common: Unify WIDs and add default test case instances
Great amount of WIDs can be common for all platforms, so moved them
into wid/<profile>.py. Left those that need to be customized.

Refactoring the creation of a list of test instances so that for
each profile there is a default instance which has been selected as
the most common from the old list.