fix GAP/BOND/BON/BV-02-C
GAP wid 108 is called twice, first time with message "Please start the
Bonding Procedure in bondable mode." and second time with "Please
configure the IUT into LE Security and start pairing process.".
As we are in bondable mode anyways, first message can just be skipped,
and we initiate pairing on second.