stack/gap: wait for multiple connections
For tests with 2 ptses we may need to verify multiple connections.
Instead of keeping record of one address let's keep list of addresses.
Modified waiting functions: waiting for "0" connections means that we
wait for any amount of connections. If we wait for more than 0,
that means that we wait until given number of connections is reached.