common: Terminate PTS and BPV after run
Sometimes PTS or BPV does not release a handle to a log file after
a test case is completed. After the last test case PTS and BPV are not
restarted, so to avoid autoptsbot failure with the exception:
"The process cannot access the file because it is being used by
another process", lets terminate them before saving the logs.