autoptsserver: Refactor stopping PTS and dongle selection
Race in PID detection could terminate the wrong autoptsserver instance.
After this refactor, we search for the PID of PTS.exe using the window
title, as the address is placed there after a dongle selection. However,
this method can fail if no dongle was selected. Then ExitPTS method of
COM server will be used. But this can fail too. Then "kill-them-all"
procedure will be berformed, but only if there is only a single instance
of autoptsserver.py on the host machine.