boards: Strict a serial number to be only numeric
The implementation of get_free_device is pretty naive and takes a first
found COM serial device if only it has any serial number. Because the
JLink serial numbers as far were only numeric and recently supported
boards has serial numbers with length >=9, we can improve the function
a little.