cron: Use 'database_file' config instead of find_latest_db
find_latest_db turned out to be not a very good solution. It works
fine if githbudrive repo is used only on one machine. Let's just use
the .db file configured with 'database_file'. The database still
can exist in the githubdrive, so let's update the repo at bot start.
The remaining issue with separate .db files will be fixed in the next
patches with some merging tool instead.