mesh: Restore IV Update Test mode after reset
IV Update Test mode flag is set in bt_mesh.flags:
https://github.com/zephyrproject-rtos/zephyr/blob/7fb2929e5ba058da82716b6555f75c112c252072/subsys/bluetooth/mesh/net.c#L248-L250
which is reset when bt_mesh_reset() is called:
https://github.com/zephyrproject-rtos/zephyr/blob/7fb2929e5ba058da82716b6555f75c112c252072/subsys/bluetooth/mesh/main.c#L341-L352

If hdl_wid_13 has been called at the start of the test and
stack.mesh.is_provisioned.data was set, btp.mesh_reset() will be called,
which will reset the flag. Therefore we need to restore it after
resetting the stack.

Signed-off-by: Pavel Vasilyev <pavel.vasilyev@nordicsemi.no>