client: LTThread: Use other lock to protect logger
Using the the logging._lock was a mistake, because it is used
internally for handlers management, not to protect a file handler
while logging. Log file handlers have their own locks. Instead of
trying to get to internal locks, we can create our own and just monky
patch the Logger._log method to make use of it globally.