zephyr: Fix GAP/SEC/AUT/BV-25-C test
This PTS test is very fragile with this test and any unexpected
pairing causes PTS to fail test (via SMP timeout).

This seems to be an issue on PTS side andthis patch is merely
a workaround for PTS issue and thus MMI handling should be reworked
when PTS fix its implementation.