Fix GAP/BOND/BON/BV-02-C
PTS seems to send WID 108 twice in a row in this test just with
different deciption. This results in calling BTP Pair command
twice. On Zephyr second call is just a NOP (since encryption
is already enabled), but in NimBLE this results in re-encrypting
link with existing LTK which confuses PTS due to unexpected HCI
event.