GAP: Add GAP pair command with sec level setting
Add a command gap_pair_with_sec_level to set a specific security level
when starting a pairing procedure.

If peer is already paired IUT is expected to enable security
(encryption) with peer. If peer is not paired IUT shall start pairing
process.
Possible values for the Security_level parameter are:
0 - Only for BR/EDR special cases, like SDP
1 - No encryption and no authentication
2 - Encryption and no authentication (no MITM)
3 - Encryption and authentication (MITM)
4 - Authenticated Secure Connections and 128-bit key

Signed-off-by: Lyle Zhu <lyle.zhu@nxp.com>