Fix GAP/SEC/AUT/BV-19-C
WID108 appears twice in a row for this test with different descritpions.
The point is to call gap_pair command only once. This was also the case
for test GAP/BOND/BON/BV-02-C, which was explained by sjanc:
"On Zephyr second call is just a NOP (since encryption is already enabled),
but in NimBLE this results in re-encrypting link with existing LTK which
confuses PTS due to unexpected HCI event."