tools: cron: Deep copy a config loaded as module
A .py file loaded as a module is cached until it is explicitly removed
from sys.modules. So subsequent attempts to load the file only return
a reference to the existing module-dictionary. So we need to deep-copy
a config.py to prevent overwriting its original values and reuse it
e.g. in subsequent PR jobs.