Log BTP headers as hex
Log BTP headers as hex instead of decimal so it is
easier to compare to the defined values in defs.py

Since the headers are using a namedtuple instead of
a class, it is not straigh forward to just update the
str or repr functions of the class.
Since printing a namedtuple uses the repr of the class,
a new function, repr_hdr, was added to format the header
as a string as hex.

Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>