bot: report: Add a note for TIMEOUT result
Add a note to report.txt, that logs may not be available for a test
case with TIMEOUT result. The TIMEOUT result means that the autopts
bot was externally terminated, because a test case took too long and
the bot failed to recover PTS. If backup mode is enabled, the bot
can be restarted to resume the test run. Since the PTS was forcefully
terminated and the faulty test case will not be rerun by the bot,
the test case logs will most likely not be available.