docs: Clean up whitespaces and unify convention used
This unify documentation format, initial idea was to use
tabs (8 spaces wide) for indentation but this seems to be causing
discrepancies in editors and with time it was not really coherent
(various places asuming 4 spaces tabs, or indentation with spaces
only).

Just unify this and use spaces for indentation (8 spaces for level).

This also fixes some convention discrepancies in BTP descriptions
so it is easier to follow.