Force pairing for CSIP tests
PTS can be sloppy with requesting pairing from IUT for CSIP tests,
being a particular problem in CSIP/CL/SP/BV-07-C. With unfortunate
timing this can lead to the IUT never receiving a pairing-request
from a lower tester, and with some controllers this happens
consistently. This commit forces the IUT to initiate pairing,
regardless of a request from the lower tester being received or not.