Fix `hdl_wid_52` decode error
The `value_read` variable is converted to a string in the if branch but
not in the else one where it's still `bytes`. The `decode` method later
called on `value_read` is trying to convert it to a string. But in the
case we were in the if branch, `value_read` is already a string. This
lead to an `AttributeError` because the `decode` method doesn't exist
for string.

Here is the error that we are getting:
```
  File "/auto-pts/autopts/wid/gatt.py", line 711, in hdl_wid_52
    value_read = value_read.decode('utf-8')
                 ^^^^^^^^^^^^^^^^^
AttributeError: 'str' object has no attribute 'decode'. Did you mean: 'encode'?
```

Introduce `value_read_str` variable to fix the issue. That variable will
be used to convert `value_read` in each branch to a string.