SM: Update wid 102
The global variable `SM_ACL_DISCONN_ROUND` is used to save the current
number times of ACL disconnection. The default value of the
`SM_ACL_DISCONN_ROUND` should be 0 when the test case is starting.

But when testing more than one case, the initialization value of the
variable SM_ACL_DISCONN_ROUND is not 0, because the previous status
is not cleared.

Add a global variable `SM_TEST_CASE_NAME` to save the test case name
corresponding to the variable `SM_ACL_DISCONN_ROUND`. When the test
case name is not same as the current test case name, clear the
variable `SM_ACL_DISCONN_ROUND` because it is a new test sequence
of new test case.

Signed-off-by: Lyle Zhu <lyle.zhu@nxp.com>