Update wid 51 to support case GAP/DISC/GENM/BV-02-C
Because the wid 52 will be called after the wid 51 called.

The case will fail if the function `hdl_wid_51` is executed.

Due to the undirectly connectable mode is required by the case,
the advertising mode is non-connectable mode set by handling of
wid 51.

Bypass the handling of wid 51 to fix the issue.