GAP: Update wid 53 to support case GAP/DISC/GENM/BV-02-C
Because the wid 52 will be called after the wid 53 called.

The case will fail if the function `hdl_wid_53` is executed.

Due to the undirectly connectable mode is required by the case,
while the advertising mode is non-connectable mode set by handling
of wid 53, bypass the handling of wid 53 to fix the issue.

Signed-off-by: Lyle Zhu <lyle.zhu@nxp.com>