GAP: Update wid 102, 103, 33, and 77
The global variable `GAP_TEST_WID_102_ROUND` is used to save the
current number times of the calling of `hdl_wid_102()`. The default
value of the `GAP_TEST_WID_102_ROUND` should be 0 when the test case
is starting.

But when testing more than one case, the initialization value of the
variable GAP_TEST_WID_102_ROUND is not 0, because the previous status
is not cleared.

Add a global variable `GAP_TEST_WID_102_TEST_CASE_NAME` to save the
test case name corresponding to the variable `GAP_TEST_WID_102_ROUND`.
When the test case name is not same as the current test case name,
clear the variable `GAP_TEST_WID_102_ROUND` because it is a new test
sequence of new test case.

Same issue also found in wid 103, 33, and 77. Apply the same changes
for the wid 103, 33, and 77.

Signed-off-by: Lyle Zhu <lyle.zhu@nxp.com>