bot: Add copy parameter
This option can be used to skip creating and using copy
of workspace file. This might be useful if you need to run tests
in automation and later use this workspace file to generate report
in PTS, or you need to repeat some test cases manually and
keep verdicts from automated test session.