wid: Fix BAP/UCL/STR/BV-530-C
This particular test case use two lower testers to set up two streams,
and as such wid 364 is called for both of them. Previously AutoPTS did
not account for separate lower testers in this wid, which caused the
second streaming event to be ignored, in turn failing the test case.
The changes in this commit fixes this issue.