wid: BAP: Refactor the ac_configs list to use classes
the ac_configs dict was an unstructured dict that required
a significant amount of comments to make sense. Refactor the dict
of unstructured data into a list of well-defined dataclass classes
and a dict copy to make lookup easier.

The changes to the code is kept to a minimal to avoid too much
complexity, but ideally the configs added to
stack.bap.ase_configs would make better use of these new classes,
and avoid any code duplication.

Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>