Timer timeout: Add logging.error when timeout occurs
When an event handler has a timeout we now log this as
an error with what information that is available.

This is a precursor to a more complete solution where
such a timeout should halt the execution of the test, instead
of continuing. The purpose of this change is to expose how many
cases of timeouts we see in the current implementation.

Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>