btp: Add new ASCS events for CIS connection state
Added 2 new events to indicate the state of the CIS
associated with an ASE, as that cannot be determined directly
from the ASE state.

The "connected" event is the only one used so far, and
is used to wait for the even when we expect the CIS to be
connected.

Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>